started writing new normalisation function which can process more cases of normalisation

+++ src/ode/ode_model.h
@@ -25,7 +25,7 @@ extern "C" {
 #define	ODE_SOLVER_MX_STEPS		5000
 	
 typedef int (*rhs_f)(realtype, N_Vector, N_Vector, void *);
-typedef int (*jac_f)(int, realtype, N_Vector, N_Vector, DlsMat, void*, N_Vector, N_Vector, N_Vector);
+typedef int (*jac_f)(long int, realtype, N_Vector, N_Vector, DlsMat, void*, N_Vector, N_Vector, N_Vector);
 typedef int (*jacp_f)(int, realtype, N_Vector, N_Vector, DlsMat, void*,	N_Vector, N_Vector, N_Vector);
 typedef int (*rhs_sens)(int, realtype, N_Vector, N_Vector, int, N_Vector, N_Vector, void *, N_Vector, N_Vector);
 typedef int (*func)(realtype, N_Vector, realtype *, void *);
